Add first draft of protocol and datatypes
I am not sure about this datatypes layout. It is based on the discovery
query used in PyHS100. They use a query that allows them to determine
the capabilities of the device without knowing exactly what it is. At
the moment I am thinking of explicitly supporting devices so that we can
verify the API usage at compile time.